parkinson_voice_detection/requirements.txt
streamlit==1.28.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
sounddevice==0.4.6
soundfile==0.12.1
praat-parselmouth==0.4.3
matplotlib==3.7.2
plotly==5.15.0